+++ src/app/components/turn-result-modal.tsx
